# swagger-document.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/server/pet/paginate": {
      "get": {
        "operationId": "PetController_paginatePets",
        "parameters": [
          {
            "name": "pet_name",
            "required": false,
            "in": "query",
            "description": "Name of the pet.",
            "example": "Luna",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pet_breed_id",
            "required": false,
            "in": "query",
            "description": "Breed of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pet_gender",
            "required": false,
            "in": "query",
            "description": "Gender of the pet.",
            "example": "Female",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pet_color",
            "required": false,
            "in": "query",
            "description": "Color of the pet.",
            "example": "Golden",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tutor_id",
            "required": false,
            "in": "query",
            "description": "ID of the tutor (owner) of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_by_created_at",
            "required": false,
            "in": "query",
            "description": "Order by created date.",
            "example": "ASC",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_by_updated_at",
            "required": false,
            "in": "query",
            "description": "Order by updated date",
            "example": "DESC",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The limit of items. If not passed the default limit will be 10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "The Current Page. If not passed the default page will be 1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pet"
        ]
      }
    },
    "/server/pet/{id}": {
      "get": {
        "operationId": "PetController_getPet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      },
      "put": {
        "operationId": "PetController_updatePet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePetDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      },
      "delete": {
        "operationId": "PetController_deletePet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      }
    },
    "/server/pet": {
      "post": {
        "operationId": "PetController_createPet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      }
    },
...
    "/server/pet-breed/paginate": {
      "get": {
        "operationId": "PetBreedController_paginateBreeds",
        "parameters": [
          {
            "name": "species_id",
            "required": false,
            "in": "query",
            "description": "Id of the species.",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "breed_name_name",
            "required": false,
            "in": "query",
            "example": "Golden",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_by_created_at",
            "required": false,
            "in": "query",
            "description": "Order by created date.",
            "example": "ASC",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_by_updated_at",
            "required": false,
            "in": "query",
            "description": "Order by updated date",
            "example": "DESC",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The limit of items. If not passed the default limit will be 10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "The Current Page. If not passed the default page will be 1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pet-breed"
        ]
      }
    },
    "/server/pet-breed/{id}": {
      "get": {
        "operationId": "PetBreedController_getPetBreedById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetBreed"
                }
              }
            }
          }
        },
        "tags": [
          "pet-breed"
        ]
      },
      "put": {
        "operationId": "PetBreedController_updateBreed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePetBreedDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetBreed"
                }
              }
            }
          }
        },
        "tags": [
          "pet-breed"
        ]
      },
      "delete": {
        "operationId": "PetBreedController_deleteBreed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetBreed"
                }
              }
            }
          }
        },
        "tags": [
          "pet-breed"
        ]
      }
    },
    "/server/pet-breed/breeds-by-species/{species_id}": {
      "get": {
        "operationId": "PetBreedController_getBreedsBySpeciesId",
        "parameters": [
          {
            "name": "species_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "pet-breed"
        ]
      }
    },
    "/server/pet-breed": {
      "post": {
        "operationId": "PetBreedController_createBreed",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetBreedDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetBreed"
                }
              }
            }
          }
        },
        "tags": [
          "pet-breed"
        ]
      }
    }
  },
  "info": {
    "title": "ikut-backend",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "hashed_password": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "user_type": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            }
          },
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "user_name",
          "hashed_password",
          "user_email",
          "user_type",
          "phone_number",
          "pets",
          "id",
          "created_at",
          "updated_at"
        ]
      },
      "PetBreed": {
        "type": "object",
        "properties": {
          "breed_name": {
            "type": "string"
          },
          "species_id": {
            "type": "string"
          },
          "species": {
            "$ref": "#/components/schemas/PetSpecies"
          },
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "breed_name",
          "species_id",
          "species",
          "id",
          "created_at",
          "updated_at"
        ]
      },
      "PetSpecies": {
        "type": "object",
        "properties": {
          "species_name": {
            "type": "string"
          },
          "breeds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetBreed"
            }
          },
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "species_name",
          "breeds",
          "id",
          "created_at",
          "updated_at"
        ]
      },
      "Pet": {
        "type": "object",
        "properties": {
          "pet_name": {
            "type": "string"
          },
          "pet_breed_id": {
            "type": "string"
          },
          "pet_species_id": {
            "type": "string"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pet_gender": {
            "type": "string"
          },
          "pet_color": {
            "type": "string"
          },
          "alergies": {
            "type": "string",
            "nullable": true
          },
          "medical_conditions": {
            "type": "string",
            "nullable": true
          },
          "current_medication": {
            "type": "string",
            "nullable": true
          },
          "pet_image_url": {
            "type": "string",
            "nullable": true
          },
          "pet_microship_id": {
            "type": "string"
          },
          "tutor_id": {
            "type": "string"
          },
          "tutor": {
            "$ref": "#/components/schemas/User"
          },
          "pet_species": {
            "$ref": "#/components/schemas/PetSpecies"
          },
          "pet_breed": {
            "$ref": "#/components/schemas/PetBreed"
          },
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "pet_name",
          "pet_breed_id",
          "pet_species_id",
          "date_of_birth",
          "pet_gender",
          "pet_color",
          "alergies",
          "medical_conditions",
          "current_medication",
          "pet_image_url",
          "pet_microship_id",
          "tutor_id",
          "tutor",
          "pet_species",
          "pet_breed",
          "id",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreatePetDTO": {
        "type": "object",
        "properties": {
          "pet_name": {
            "type": "string",
            "description": "Name of the pet.",
            "example": "Luna"
          },
          "pet_breed_id": {
            "type": "string",
            "description": "Breed of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "description": "Date of birth of the pet.",
            "example": "2020-01-15"
          },
          "pet_gender": {
            "type": "string",
            "description": "Gender of the pet.",
            "example": "Female"
          },
          "pet_color": {
            "type": "string",
            "description": "Color of the pet.",
            "example": "Golden"
          },
          "alergies": {
            "type": "string",
            "description": "Any allergies the pet may have.",
            "example": "Pollen"
          },
          "medical_conditions": {
            "type": "string",
            "description": "Any medical conditions the pet may have.",
            "example": "Hip dysplasia"
          },
          "current_medication": {
            "type": "string",
            "description": "Current medication the pet is taking.",
            "example": "None"
          },
          "pet_image_url": {
            "type": "string",
            "description": "URL of the pet's image.",
            "example": "https://example.com/pet.jpg"
          },
          "pet_microship_id": {
            "type": "string",
            "description": "Microchip ID of the pet.",
            "example": "A123456789"
          },
          "tutor_id": {
            "type": "string",
            "description": "ID of the tutor (owner) of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "pet_species_id": {
            "type": "string",
            "description": "ID of the species the pet belongs to.",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        },
        "required": [
          "pet_name",
          "pet_breed_id",
          "pet_gender",
          "pet_color",
          "tutor_id",
          "pet_species_id"
        ]
      },
      "UpdatePetDTO": {
        "type": "object",
        "properties": {
          "pet_name": {
            "type": "string",
            "description": "Name of the pet.",
            "example": "Luna"
          },
          "pet_breed_id": {
            "type": "string",
            "description": "Breed of the pet.",
            "example": "Labrador Retriever"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "description": "Date of birth of the pet.",
            "example": "2020-01-15"
          },
          "pet_gender": {
            "type": "string",
            "description": "Gender of the pet.",
            "example": "Female"
          },
          "pet_color": {
            "type": "string",
            "description": "Color of the pet.",
            "example": "Golden"
          },
          "alergies": {
            "type": "string",
            "description": "Any allergies the pet may have.",
            "example": "Pollen"
          },
          "medical_conditions": {
            "type": "string",
            "description": "Any medical conditions the pet may have.",
            "example": "Hip dysplasia"
          },
          "current_medication": {
            "type": "string",
            "description": "Current medication the pet is taking.",
            "example": "None"
          },
          "pet_image_url": {
            "type": "string",
            "description": "URL of the pet's image.",
            "example": "https://example.com/pet.jpg"
          },
          "pet_microship_id": {
            "type": "string",
            "description": "Microchip ID of the pet.",
            "example": "A123456789"
          },
          "tutor_id": {
            "type": "string",
            "description": "ID of the tutor (owner) of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "pet_species_id": {
            "type": "string",
            "description": "ID of the species the pet belongs to.",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        },
        "required": [
          "tutor_id",
          "pet_species_id"
        ]
      },
...
      "CreatePetBreedDTO": {
        "type": "object",
        "properties": {
          "breed_name": {
            "type": "string",
            "description": "Name of the pet."
          },
          "species_id": {
            "type": "string",
            "description": "Id of the species.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "breed_name",
          "species_id"
        ]
      },
      "UpdatePetBreedDTO": {
        "type": "object",
        "properties": {
          "breed_name": {
            "type": "string",
            "description": "Name of the pet.",
            "example": "Golden Retriever"
          },
          "species_id": {
            "type": "string",
            "description": "Name of the pet.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}